[Entities] Added RenderJobLog entity.
